Add files via upload
This repository contains insights and conclusions derived from analyzing the data set of the company "PLUM". The data set provided information on the success rate of onboarding customers, the percentage of customers opting for different policies, the distribution of orders among different insurers, and the performance of individual team members in processing orders.

Based on the analysis, it was found that fresh orders are significantly higher than renewals and 93% of customers opted for GMC policies. The top 5 insurers handled 87% of the customers. However, the majority of the orders are still in processing stage.

The onboarding team can significantly increase the success rate by fastening the collection of policies and invoices. Additionally, by concentrating more on orders from the top 5 insurers in processing stage and putting more efforts, the growth of the company can be increased efficiently.

This repository provides valuable insights and conclusions that can help the company "PLUM" improve its onboarding process and achieve better efficiency in servicing its customers.