Add files via upload
The comprehensive exploratory data analysis (EDA) report for the "PLUM" company, which includes in-depth analysis and visualizations of the company's customer data. The HTML file provides a detailed overview of the data insights and conclusions, as well as key observations such as the percentage of customers opting for GMC policies and the top 5 insurers handling 87% of the customers.